(get) item

+++ src/main/java/com/green/gallery_jwt_jpa/application/item/ItemController.java
@@ -29,7 +29,7 @@ public class ItemController {
     }
 
     @GetMapping
-    public ResponseEntity<?> readAll(@RequestParam(name = "id", required = false) ArrayList<Integer> ids) {
+    public ResponseEntity<?> readAll(@RequestParam(name = "id", required = false) ArrayList<Long> ids) {
         log.info("ids: {}", ids);
         List<ItemGetRes> items = itemService.findAll(ids);
         return ResponseEntity.ok(items);
